backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
cohere==5.5.3
pyjwt[crypto]==2.8.0
httpx==0.25.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.13.1